backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
celery==5.3.6
redis==5.0.3
pydantic==2.6.4
pydantic-settings==2.2.1
google-genai>=1.0.0
PyMuPDF==1.24.1
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.2.0
python-multipart==0.0.9
python-dotenv==1.0.1
